refactor: remove extra sermons component
