+++ client/apps/user-portal/src/app/routing.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { SermonsComponent } from './sermons/sermons.component';
-import { SermonViewComponent } from './sermons/sermon-view/sermon-view.component';
+import { SermonsComponent } from '@sb/sermons';
+import { SermonViewComponent } from './sermon-view/sermon-view.component';
 
 const routes: Routes = [
   {
