+++ client/apps/user-portal/src/app/sermon-view/sermon-view.component.ts
