--- client/apps/user-portal/src/app/sermons/sermons.component.html
@@ -1,5 +1,0 @@
-<div class="d-flex justify-content-center mt-3">
-  <sb-searchbar></sb-searchbar>
-</div>
-
-<sb-sermon-list></sb-sermon-list>
