--- client/apps/user-portal/src/app/sermons/sermons.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-sermons',
-  templateUrl: './sermons.component.html',
-  styleUrls: ['./sermons.component.scss']
-})
-export class SermonsComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
